requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
Pillow
transliterate
drf-yasg
django-rest-passwordreset
